+++ apps/repository/js/MessageBanner.tsx
@@ -112,7 +112,12 @@ const MESSAGES: Message[] = [
     {
         id: 'opencollective-donate',
         element: <div>Polar needs your help to continue future development! <a href="https://opencollective.com/polar-bookshelf/donate">Donate to Support Polar</a></div>
-    }
+    },
+    {
+        id: 'polar-feedback-2018-11',
+        element: <div>Could you take 1 minute and <a href="https://kevinburton1.typeform.com/to/SsuiyB">complete a survey</a> about your usage of Polar?</div>
+    },
+
     // {
     //     id: 'opencollective-donate-button',
     //     element: <div>